Now saves file...
The program now saves the vertretungsplan to a specified backup path.
So, now the program can recognise if a file was already polled,
converted, sent,... and doesn't do those steps again. It just compares
the file from the internet with the stored file and checks if the
vertretungsplan updated.

Updated .gitignore