Don't shows unnecessary lessons with no vertretung
If there are 5 lessons (1.Std.,2.Std.,...) and only 1.Std. and 3.Std.
have content. The program will now only show 1.Std.,2.Std. and 3.Std. in
it's output. 2.Std. has no content, but it is shown that no
misunderstandings occur, when looking at the output. But 4.Std. and
5.Std. are NOW not anymore in the ouput because they aren't needed
(unnecessary). But again if now another lesson (6.Std.) would be added
4.Std. and 5.Std. would be shown so that no misunderstandings occur. So
the program looks in reversed order and it sorts out all lessons with no
vertretung UNTIL it finds any lesson with vertretung.