Fixed encoding bug.
This bug fixes an issue where the vertretungsplan would be sent every time checked due to an encoding issue.

The website would return \r\n as newline characters, the stored file (which is needed for comparison, if the plan updated) would contain the \n newline character, also if this stored file would contain \r\n characters they would be replaced with \n. This results into the program continuously thinking that it found a new vertretungsplan.